+++ CHANGELOG.md
@@ -35,3 +35,4 @@
 - update version number to 1.4.5
 - Add Redstone, Glowstone & Ender Metals
 - Add Alloy recipes for Red Alloy, Lumium, Signalum & Enderium
+- Improve look of Redstone, Glowstone & Ender Ingot Textures + Fluid jsons